rollback the changes which code style can not pass

+++ config/src/main/java/com/alibaba/nacos/config/server/service/capacity/GroupCapacityPersistService.java
@@ -154,7 +154,7 @@ public class GroupCapacityPersistService {
         if (result == null) {
             throw new IllegalArgumentException("configInfoCount error");
         }
-        return result;
+        return result.intValue();
     }
     
     /**
